Reorganize contrib op schemas (#10494)
